[AMQCPP-56] - make buffer lengths use size_t for 64-bit operating systems
git-svn-id: https://svn.apache.org/repos/asf/incubator/activemq/activemq-cpp/trunk/activemq-cpp@500349 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/activemq/connector/openwire/utils/OpenwireStringSupport.cpp
@@ -131,7 +131,7 @@ void OpenwireStringSupport::writeString( io::DataOutputStream& dataOut,
                     __LINE__,
                     ( std::string( "OpenwireStringSupport::writeString - Cannot marshall " ) + 
                     "string longer than: 65536 characters, supplied steing was: " +
-                    Integer::toString( str->size() ) + " characters long." ).c_str() );
+                    Integer::toString( (int)str->size() ) + " characters long." ).c_str() );
             }
             
             //short strlen = (short)str->size();
